Rename the metadata envelope generator plugin.
