+++ lib/preparermd/plugins/metadata_envelopes.rb
